+++ lib/irccd/hook.h
@@ -27,8 +27,6 @@
  * events but are more limited.
  */
 
-#include <sys/types.h>
-
 #include <ev.h>
 
 #if defined(__cplusplus)
@@ -36,54 +34,6 @@ extern "C" {
 #endif
 
 struct irc_event;
-
-/**
- * \brief Hook active children.
- */
-struct irc_hook_child {
-	/**
-	 * (read-only)
-	 *
-	 * Process PID.
-	 */
-	pid_t pid;
-
-	/**
-	 * (read-only)
-	 *
-	 * Process watcher for reaping.
-	 */
-	struct ev_child child;
-
-	/**
-	 * (read-only)
-	 *
-	 * Timer in case the process would not stop.
-	 */
-	struct ev_timer timer;
-
-	/**
-	 * (read-only, borrowed)
-	 *
-	 * Parent hook in which this process belong to.
-	 */
-	struct irc_hook *parent;
-
-	/**
-	 * \cond IRC_PRIVATE
-	 */
-
-	/**
-	 * (private)
-	 *
-	 * Next child in the linked list.
-	 */
-	struct irc_hook_child *next;
-
-	/**
-	 * \endcond IRC_PRIVATE
-	 */
-};
 
 /**
  * \brief IRC event hook.
@@ -107,19 +57,7 @@ struct irc_hook {
 	 * \cond IRC_PRIVATE
 	 */
 
-	/**
-	 * (private)
-	 *
-	 * Next hook in the linked list.
-	 */
 	struct irc_hook *next;
-
-	/**
-	 * (private)
-	 *
-	 * Linked list of active processes.
-	 */
-	struct irc_hook_child *children;
 
 	/**
 	 * \endcond IRC_PRIVATE
@@ -129,8 +67,6 @@ struct irc_hook {
 /**
  * Create a new hook.
  *
- * \pre name != NULL
- * \pre path != NULL
  * \param name the hook name
  * \param path path to the hook script/executable
  * \return a new hook
@@ -139,27 +75,16 @@ struct irc_hook *
 irc_hook_new(const char *name, const char *path);
 
 /**
- * Invoke the hook by spawning a new child with the given event.
+ * Invoke the hook and wait for its termination.
  *
- * The child process runs independently and will be reaped automatically when it
- * exits.
- *
- * \pre hook != NULL
- * \pre ev != NULL
- * \param hook the hook
- * \param ev the event to pass to the hook child
+ * \note This function will block until the hook terminated.
+ * \param ev the event to pass to the hook child process
  */
 void
 irc_hook_invoke(struct irc_hook *hook, const struct irc_event *ev);
 
 /**
- * Destroy the hook and possibly the active children if any.
- *
- * This function blocks if there is one or more active children, calling
- * `ev_run` until they are entirely destroyed.
- *
- * The hook receive a SIGTERM signal first and if they would not exit after a
- * short amount of time, SIGKILL is then raise.
+ * Destroy the hook.
  */
 void
 irc_hook_free(struct irc_hook *hook);